+++ service/controller/resource/s3object/update.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/operatorkit/v2/pkg/resource/crud"
+	"github.com/giantswarm/operatorkit/v4/pkg/resource/crud"
 
 	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
 )